Model/changeDrawingFileCase.php
adjust the drwaing filename format so that it has a consistency format.

Model/inserDrawingPath.php
analysis the drawing number of a record and generate the corresponding filepath for the record.


The above program only handle the file origanaled form FOCUS.
Once the Database is used, the location information may be change to a different format.